RAG/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.4.0
attrs==25.1.0
banks==2.2.0
beautifulsoup4==4.12.3
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
et_xmlfile==2.0.0
fastapi==0.120.2
filetype==1.2.0
frozenlist==1.4.1
fsspec==2025.9.0
greenlet==3.0.3
griffe==1.4.0
h11==0.14.0
httpcore==1.0.5
httpx==0.28.1
idna==3.7
Jinja2==3.1.4
jiter==0.7.0
joblib==1.4.2
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index-cli==0.5.3
llama-index-core==0.14.6
llama-index-embeddings-openai==0.5.1
llama-index-indices-managed-llama-cloud==0.9.4
llama-index-instrumentation==0.4.2
llama-index-llms-ollama==0.9.0
llama-index-llms-openai==0.6.6
llama-index-readers-file==0.5.4
llama-index-readers-llama-parse==0.5.1
llama-index-workflows==2.9.1
llama-parse==0.6.54
MarkupSafe==3.0.2
marshmallow==3.26.1
multidict==6.0.4
mypy_extensions==1.0.0
networkx==3.3
nltk==3.9.1
numpy==2.2.1
ollama==0.6.1
openai==1.109.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pillow==10.4.0
platformdirs==4.3.7
propcache==0.4.1
pydantic==2.11.5
pydantic_core==2.33.2
pypdf==6.1.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2025.10.23
requests==2.32.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.49.1
striprtf==0.0.26
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.66.5
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.2.3
uvicorn==0.32.0
wrapt==1.16.0
yarl==1.18.3
